# frontend/public/contracts/openapi.runtime-supplement.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Bridge AI OS Runtime Supplement",
    "version": "1.0.0",
    "description": "Supplemental live endpoints present in code but not in canonical openapi.json."
  },
  "paths": {
...
    "/api/payments/webhook/{rail}": {
      "post": {
        "summary": "Webhook Generic",
        "description": "Generic webhook for any internal project.\nCalled by AOE, Supaco, Taurus, etc. to route revenue into the unified treasury.\nBody: { amount, currency, source_project, type, customer, reference, meta }",
        "operationId": "webhook_generic_api_payments_webhook__rail__post",
        "parameters": [
          {
            "name": "rail",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Rail"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/treasury/collect": {
      "post": {
        "summary": "Treasury Collect",
        "description": "Collect revenue from any project into the unified treasury.\nRequired: amount (float)\nOptional: currency (default BRDG), source_project, method, type, meta",
        "operationId": "treasury_collect_api_treasury_collect_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Payload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/treasury/disburse": {
      "post": {
        "summary": "Treasury Disburse",
        "description": "Record a treasury disbursement from a bucket.\nRequired: bucket (ubi|treasury|ops|founder), amount, destination\nOptional: authorized_by",
        "operationId": "treasury_disburse_api_treasury_disburse_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Payload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/treasury/ledger": {
      "get": {
        "summary": "Treasury Ledger",
        "description": "Recent treasury transactions (most recent first). Max 200.",
        "operationId": "treasury_ledger_api_treasury_ledger_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}